+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Screen reader support for hints
 - Alt text to be hidden for decorative images
-- <MaterialIcon> element
+- `<MaterialIcon />` element
+- `<IconComponent />` component
 
 ### Changed